Fixed inputting error in test

+++ src/test/scala/AlgebraTest.scala
@@ -140,7 +140,7 @@ class IntegerTest extends FunSuite {
         assert(Integer(7).lambda(2) === Integer(21))
         assert(Integer(4).lambda(3) === Integer(4))
         assert(Integer(5).lambda(1) === Integer(5))
-        assert(Integer(40).lambda(30) === Integer(84740528))
+        assert(Integer(40).lambda(30) === Integer(847660528))
         assert(Integer(45).lambda(39) === Integer(8145060))
     }
 }
